Added support for UNNotificationRequest
Removed the interval option and replaced it with repeats which acts as a boolean, and you can now choose the triggertype by defining trigger : 'timeinterval'. Defaults to UNCalendarNotificationTrigger. if timeinterval is chosen, the at parameter will have to be amount of seconds from the scheduling time until the notification should appear